IMAGES-TAKEN-FROM-THE-COMPUTER-CAMERA-ARE-PROCESSED-SLEEPING-DRIVER-STIMULATION-SYSTEM
The system monitors the driver’s condition through a night vision mini camera placed
on the upper part of the vehicle’s dashboard. Evaluating the driver’s condition based
on predefined criteria, the system activates warnings when necessary. This ensures that
the driver is informed before transitioning into a drowsy state.